Refactor expense sheet form
