+++ frontend/src/views/expense_sheets/expense_sheet_form/segments/expenseSheetFormSegment.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { SolidHorizontalRow } from '../../../../layout/SolidHorizontalRow';
+
+export const expenseSheetFormSegment: <T>(Component: React.ComponentType<T>) => (props: T) => React.ReactElement = Component => props => (
+  <>
+    <Component {...props}/>
+    <SolidHorizontalRow/>
+  </>
+);
