+++ frontend/src/views/expense_sheets/expense_sheet_form/segments/index.ts
@@ -1,0 +1,21 @@
+import { AbsolvedDaysBreakdownSegment } from './AbsolvedDaysBreakdownSegment';
+import { ClothingExpensesSegment } from './ClothingExpensesSegment';
+import { CompanyHolidaysSegment } from './CompanyHolidaysSegment';
+import { DrivingExpensesSegment } from './DrivingExpensesSegment';
+import { ExtraordinaryExpensesSegment } from './ExtraordinaryExpensesSegment';
+import { FooterSegment } from './FooterSegment';
+import { GeneralSegment } from './GeneralSegment';
+import { PaidVacationSegment } from './PaidVacationSegmet';
+import { UnpaidVacationSegment } from './UnpaidVacationSegmet';
+
+export {
+  AbsolvedDaysBreakdownSegment,
+  ClothingExpensesSegment,
+  CompanyHolidaysSegment,
+  DrivingExpensesSegment,
+  ExtraordinaryExpensesSegment,
+  FooterSegment,
+  GeneralSegment,
+  PaidVacationSegment,
+  UnpaidVacationSegment,
+};
